add light project view
